feat : adjustments

+++ rrrbazar-api/src/models/AdminTransaction.ts
@@ -31,7 +31,7 @@ export default (sequelize: Sequelize) => {
             allowNull: true,
         },
         amount: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(10, 2),
             allowNull: true,
         },
         completed_by: {
